backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
segmentation-models-pytorch
Pillow
